ice: prevent infinite loop on unclean shutdown
If the remote side leaves without terminating, then an exception is
raised and the connection is closed.

The function `query_consent()` makes a protocol request, which can
fail in a number of ways including timing out.

When this happens, `query_consent()` calls `self.close()`.

Unfortunately, `self.close()` waits for the
`self._query_consent_handle` call to finish, which will never happen
because it is waiting on `self.close()`.

To work around this issue, null-out `self._query_consent_handle`
prior to calling `self.close()`.

Signed-off-by: Sean Cross <sean@xobs.io>