Switch from `netifaces` to `ifaddr` (fixes: #62)
The `netifaces` package is no longer maintained, so no new wheels are
being published. As it is not a pure Python package, this means that
installing `aioice` would require users to compile `netifaces`
themselves which is not trivial on some platforms.

We use the `ifaddr` package instead, which is a pure Python package.